ml/pyproject.toml
[project]
name = "aura-ml"
version = "0.0.0"
description = "Aura ML pipeline: Qwen-Image-Edit + Qwen3.5-9B prompt expander + procedure LoRAs"
requires-python = ">=3.11,<3.13"

dependencies = [
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.46.0",
    "diffusers>=0.32.0",
    "peft>=0.14.0",
    "accelerate>=1.2.0",
    "bitsandbytes>=0.45.0",
    "safetensors>=0.4.5",
    "huggingface-hub>=0.27.0",
    "pillow>=11.0.0",
    "numpy>=2.0.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
eval = [
    "lpips>=0.1.4",
    "open-clip-torch>=2.30.0",
    "insightface>=0.7.3",
    "onnxruntime-gpu>=1.20.0",
    "scikit-image>=0.24.0",
]
demo = [
    "gradio>=5.0.0",
    "openai-whisper>=20240930",
]
dev = [
    "ruff>=0.7.0",
    "ipython>=8.30.0",
    "ipykernel>=6.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aura_ml"]

# --- uv: pull torch from the cu128 (Blackwell) wheel index ---
[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]
